Move slack controller unnder bot/ directory
